Make VSCode extension kogito exclusive to BPMN
