+++ packages/vscode-extension-pack-kogito-bpmn/.vscodeignore
