+++ packages/vscode-extension-pack-kogito-bpmn/jest.config.js
