+++ packages/vscode-extension-pack-kogito-bpmn/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "appformer-js-vscode-extension-pack-kogito",
-  "displayName": "KIE :: Kogito Editors",
+  "name": "appformer-js-vscode-extension-pack-kogito-bpmn",
+  "displayName": "KIE :: Kogito Editors :: BPMN",
   "description": "",
   "version": "0.1.0",
   "license": "Apache-2.0",
@@ -22,16 +22,6 @@
   "contributes": {
     "languages": [
       {
-        "id": "dmn",
-        "extensions": [
-          ".dmn"
-        ],
-        "aliases": [
-          "dmn",
-          "Decision Model Notation"
-        ]
-      },
-      {
         "id": "bpmn",
         "extensions": [
           ".bpmn",
@@ -46,7 +36,7 @@
     ]
   },
   "scripts": {
-    "pack-extension": "vsce package --yarn -o ./dist/kiegroup_kogito_editors_$npm_package_version.vsix",
+    "pack-extension": "vsce package --yarn -o ./dist/kiegroup_kogito_editors_bpmn_$npm_package_version.vsix",
     "compile": "webpack",
     "watch": "webpack",
     "test": "jest",
