+++ packages/vscode-extension-pack-kogito-bpmn/src/__tests__/DummyEditor.tsx
