+++ packages/vscode-extension-pack-kogito-bpmn/src/common/KogitoLanguageData.ts
