+++ packages/vscode-extension-pack-kogito-bpmn/src/resources/YouTube_loading_symbol_2_(stable).gif
