+++ packages/vscode-extension-pack-kogito-bpmn/src/webview/gwt/GwtEditorWrapperFactory.ts
