+++ packages/vscode-extension-pack-kogito-bpmn/tests-it/test-workspace/example.txt
