+++ packages/vscode-extension-pack-kogito-bpmn/tests-it/tsconfig.json
