+++ packages/vscode-extension-pack-kogito-bpmn/webpack.config.js
@@ -75,11 +75,6 @@ module.exports = [
     entry: {
       "webview/index": "./src/webview/index.ts"
     },
-    plugins: [
-      new CopyWebpackPlugin([
-        { from: "../unpacked-gwt-editors/dmn", to: "webview/editors/dmn" },
-        { from: "../unpacked-gwt-editors/bpmn", to: "webview/editors/bpmn" }
-      ])
-    ]
+    plugins: [new CopyWebpackPlugin([{ from: "../unpacked-gwt-editors/bpmn", to: "webview/editors/bpmn" }])]
   }
 ];
